+++ catsoop/scripts/util/10_to_11/convert_checker.py
@@ -9,7 +9,7 @@ sys.path.append(catsoop_root)
 import catsoop.cslog as cslog
 import catsoop.base_context as bc
 
-results = os.path.join(bc.cs_data_root, "__LOGS__", "_checker", "results")
+results = os.path.join(bc.cs_data_root, "_logs", "_checker", "results")
 for fname in sorted(os.listdir(results)):
     if len(fname) == 1:
         continue